Fix await inside while loop
