+++ js/index.js
@@ -1,11 +1,10 @@
 import utils from './utils.js';
-import { grabRandomUser, populateUsers } from './randomUser.js';
+import { grabRandomUser } from './randomUser.js';
 
 const burgerMenuIcon = utils.qs('header img');
 const burgerMenuIconClose = utils.qs('.mobile-menu img');
 const burgerMenu = utils.qs('.mobile-menu');
 const burgerMenuBackground = utils.qs('div', burgerMenu);
-const menuList = utils.qs('ul.desktop-visible');
 const header = utils.qs('header');
 const scrollToTop = utils.qs('#scrollToTop');
 
@@ -14,7 +13,7 @@ const closeMenu = () => {
 
   burgerMenu.addEventListener(
     'transitionend',
-    (e) => {
+    () => {
       ul.remove();
       ul.classList.add('desktop-visible');
       utils.qs('header').appendChild(ul);
@@ -58,22 +57,7 @@ window.addEventListener('resize', () => {
   }
 });
 
-const grabSpeakers = async () => {
-  populateUsers(await grabRandomUser());
-
-  const button = utils.qs('#teachers button');
-
-  button.textContent = 'more';
-  button.appendChild(
-    utils.createElement({
-      tagName: 'img',
-      src: './img/burger/arrow_down.png',
-      alt: 'More speakers',
-    })
-  );
-};
-
-grabSpeakers();
+grabRandomUser();
 
 window.addEventListener('scroll', () => {
   if (header.getBoundingClientRect().bottom < 0) {
